# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API for Icloud",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get all documents",
                "operationId": "cf58d1781992e8292fd9914b24487fdc",
                "parameters": [
                    {
                        "name": "filter[relevance]",
                        "in": "query",
                        "description": "Filter by relevance. Posible filters are 'alta' 'media' and 'baja'",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[starts_before]",
                        "in": "query",
                        "description": "Filter by aproval date. Date must be in YYYY-MM-DD format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[starts_after]",
                        "in": "query",
                        "description": "Filter by aproval date. Date must be in YYYY-MM-DD format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting documents by id, name, relevance or approval date. Adding - indicates descending order of results. Ex.: -id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of documents"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Create a new document",
                "operationId": "52f8c51858a6aacad40c22d0233e4f02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the document",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the document",
                                        "type": "string"
                                    },
                                    "relevance": {
                                        "description": "Relevance of the document",
                                        "type": "string",
                                        "enum": [
                                            "alta",
                                            "media",
                                            "baja"
                                        ]
                                    },
                                    "approval_date": {
                                        "description": "Approval date of the document",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "upload_date": {
                                        "description": "Upload date of the document",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "pdf_path": {
                                        "description": "Path to the PDF file",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/documents/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a document by ID",
                "operationId": "9cf5289c9ad15ddb6da247477998f54b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/DocumentResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Docuement retrieve successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Documents"
                ],
                "summary": "Update an existing document",
                "operationId": "8340a1a99f82c944b6c7b45b0794505f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Document updated successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Delete a document by ID",
                "operationId": "797f07f3ce355da8e70bf49bb13caa6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DocumentResource": {
                "description": "A single document resource with metadata",
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 100
                            },
                            "name": {
                                "type": "string",
                                "example": "Amet debitis qui enim."
                            },
                            "description": {
                                "type": "string",
                                "example": "Placeat eveniet et facilis similique. Quo quas quo cum dignissimos possimus et dolores."
                            },
                            "relevance": {
                                "type": "string",
                                "example": "media"
                            },
                            "approval_date": {
                                "type": "string",
                                "format": "date-time",
                                "example": "05-10-2022 00:00:00"
                            },
                            "upload_date": {
                                "type": "string",
                                "format": "date-time",
                                "example": "30-07-2024 00:00:00"
                            },
                            "pdf_path": {
                                "type": "string",
                                "example": "documents/slHq5FuhBb.pdf"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "!IMPORTANT dont forget the prefix Bearer! ej: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}